Adds a bitset encoded BBMC sequential Max Clique solver
TL;DR this bitset encoding BBMC implementation is significantly slower
than the current sequential Max Clique solver in Solvers.Sequential,
but has the potential to be parallelised.

This is a Haskell port of the BBMC max clique algorithm:
http://www.dcs.gla.ac.uk/~pat/maxClique/distribution/java-20120710/BBMC.java

Described in Section 2.3 "Bitset Encodings", of:"Multi-Threaded Maximum Clique".
Ciaran McCreesh, University of Glasgow. March 2013.
http://www.dcs.gla.ac.uk/~pat/4yProjects/HallOfFame/reports2013/McCreeshC.pdf

The algorithm is "Algorithm 3: bmcsa, a bitset encoding of mcsa".  The
heavy lifting of this implementation corresponds to the `colourise`
function as it is written in the algorithm. The inner `while` loop
from line 29 correspondds to bbColour.goInner in this Haskell
module. It is here that the runtime spends most of its time.

This module is unfortunately a lot significantly slower than the other
sequential algorithm in the 'Sequential' module and the reference Java
code BBMC.java. The bottleneck is bbColour.goInner. This bitset
version though can be parallelised, see Chapter 3 of the above PDF
document.