Don't reverse the task list for Broadcast skeleton
I'm not sure why I did this in the first place since this seems to go
against the heuristics.

I guess it means that those with good heuristics are likely to be stolen
first rather than ran locally, but now we have a priority queue (but a
single priority) this is not the case.